Store chat id instead of whole chat object on initial chat create
If the user already has a chat, their most recent one is loaded when the chat widget initializes. If they don't, it's created when they try to send their first message. In that latter case, the library code was incorrectly storing the ID of the newly created chat, causing subsequent messages to fail. (It was storing the whole chat where the id belongs.)